+++ docs/changelog.md
@@ -13,6 +13,7 @@ category: Administration
 
 * feat: add support for Jira 8.14.1 and latest enterprise patch releases
 * feat: fist version of the getting started page and better exception handling
+* test: add tests for getting started page
 
 ### [21.01.0] - 2021-01-08